src/test/java/in/amal/rpg/gameofstones/view/BasicViewTest.java
package in.amal.rpg.gameofstones.view;

import static org.junit.Assert.assertEquals;

import java.util.Optional;

import org.junit.Test;

import in.amal.rpg.gameofstones.BaseTest;

public class BasicViewTest extends BaseTest {

	private BasicView view = new BasicView();

	@Test
	public void defaultInputTest() {
		setUserInput("s5".getBytes());
		int option = view.readIntUserInput(Optional.of(3), new int[] { 1, 2, 3 });
		assertEquals(option, 3);
	}

	@Test
	public void noDefaultValueTest() {
		setUserInput(("1".getBytes()));
		int option = view.readIntUserInput(Optional.empty(), new int[] { 1, 2, 3 });
		assertEquals(option, 1);
	}

	@Test
	public void noDefaultValueWithWrongUserInputTest() {
		setUserInput("asdsa\n1".getBytes());
		int option = view.readIntUserInput(Optional.empty(), new int[] { 1, 2, 3 });
		assertEquals(option, 1);
	}

	@Test
	public void validInputTest() {
		setUserInput("1".getBytes());
		int option = view.readIntUserInput(Optional.of(3), new int[] { 1, 2, 3 });
		assertEquals(option, 1);
	}

}
